requirements.txt
anyio==4.11.0
astor==0.8.1
attrdict==2.0.1
babel==2.17.0
bce-python-sdk==0.9.52
beautifulsoup4==4.14.2
blinker==1.9.0
cachetools==6.2.2
Camelot==12.6.29
certifi==2025.11.12
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.4
click==8.1.8
colorama==0.4.6
contourpy==1.3.0
cryptography==46.0.3
cssselect==1.3.0
cssutils==2.11.1
cycler==0.12.1
Cython==3.2.1
decorator==5.2.1
python-dotenv==1.2.1
Elixir==0.7.1
et_xmlfile==2.0.0
exceptiongroup==1.3.0
fire==0.7.1
Flask==3.1.2
flask-babel==4.0.0
fonttools==4.60.1
future==1.0.0
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
ImageIO==2.37.2
imgaug==0.4.0
importlib_metadata==8.7.0
importlib_resources==6.5.2
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.2
kiwisolver==1.4.7
lazy_loader==0.4
lmdb==1.7.5
lxml==6.0.2
MarkupSafe==3.0.3
matplotlib==3.9.4
more-itertools==10.8.0
networkx==3.2.1
numpy==1.23.5
opencv-contrib-python==4.6.0.66
opencv-python==4.6.0.66
opencv-python-headless==4.12.0.88
openpyxl==3.1.5
opt-einsum==3.3.0
packaging==25.0
paddleocr==2.7.0.3
paddlepaddle==2.6.1
pandas==2.3.3
pbr==7.0.3
pdf2docx==0.5.8
pdf2image==1.17.0
pdfminer.six==20251107
pdfplumber==0.11.8
pillow==11.3.0
premailer==3.10.0
protobuf==3.20.2
psutil==7.1.3
psycopg==3.2.12
pyclipper==1.3.0.post6
pycparser==2.23
pycryptodome==3.23.0
PyMuPDF==1.20.2
pyparsing==3.2.5
pypdfium2==5.0.0
PySide6==6.10.0
PySide6_Addons==6.10.0
PySide6_Essentials==6.10.0
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-docx==1.2.0
pytz==2025.2
PyYAML==6.0.3
RapidFuzz==3.13.0
rarfile==4.2
requests==2.32.5
scikit-image==0.24.0
scikit-learn==1.6.1
scipy==1.13.1
setuptools==58.1.0
shapely==2.0.7
shiboken6==6.10.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
SQLAlchemy==2.0.44
sqlalchemy-migrate==0.11.0
sqlparse==0.5.3
Tempita==0.6.0
termcolor==3.1.0
threadpoolctl==3.6.0
tifffile==2024.8.30
tqdm==4.67.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
visualdl==2.5.3
Werkzeug==3.1.3
xlrd==0.7.1
xlwt==0.7.2
zipp==3.23.0